+++ packages/hash/frontend/src/components/hooks/useCreatePage.ts
@@ -11,7 +11,7 @@ import {
 } from "../../graphql/apiTypes.gen";
 import { getAccountPagesTree } from "../../graphql/queries/account.queries";
 import { createPage } from "../../graphql/queries/page.queries";
-import { useWorkspaceShortnameByEntityUuid } from "./use-get-workspace-shortname-by-entity-uuid";
+import { useWorkspaceShortnameByEntityUuid } from "./use-workspace-shortname-by-entity-uuid";
 
 export const useCreatePage = (ownedById: string) => {
   const router = useRouter();
