+++ packages/hash/frontend/src/components/hooks/useCreateSubPage.ts
@@ -13,7 +13,7 @@ import {
 } from "../../graphql/apiTypes.gen";
 import { getAccountPagesTree } from "../../graphql/queries/account.queries";
 import { createPage, setParentPage } from "../../graphql/queries/page.queries";
-import { useWorkspaceShortnameByEntityUuid } from "./use-get-workspace-shortname-by-entity-uuid";
+import { useWorkspaceShortnameByEntityUuid } from "./use-workspace-shortname-by-entity-uuid";
 
 export const useCreateSubPage = (ownedById: string) => {
   const router = useRouter();
